migration/requirements.txt
azure-identity>=1.16.0
azure-kusto-data>=4.6.0
azure-storage-blob>=12.20.0
psycopg2-binary>=2.9.9
